+++ packages/global-styles-ui/src/root-menu.tsx
@@ -4,7 +4,6 @@ import {
 	typography,
 	color,
 	layout,
-	postCommentsForm,
 	shadow as shadowIcon,
 } from '@wordpress/icons';
 import { __ } from '@wordpress/i18n';
@@ -37,7 +36,6 @@ function RootMenu() {
 	const hasShadowPanel = true; // Same as Gutenberg
 	const hasDimensionsPanel = useHasDimensionsPanel( settings );
 	const hasLayoutPanel = hasDimensionsPanel;
-	const hasFormsPanel = hasTypographyPanel || hasColorPanel;
 
 	return (
 		<>
@@ -53,14 +51,6 @@ function RootMenu() {
 				{ hasColorPanel && (
 					<NavigationButtonAsItem icon={ color } path="/colors">
 						{ __( 'Colors' ) }
-					</NavigationButtonAsItem>
-				) }
-				{ hasFormsPanel && (
-					<NavigationButtonAsItem
-						icon={ postCommentsForm }
-						path="/forms"
-					>
-						{ __( 'Forms' ) }
 					</NavigationButtonAsItem>
 				) }
 				{ hasBackgroundPanel && (